Adjust my layers to have alternate hand modifiers
I did this because I found it very awkward to use a modifier on the same
hand as the action when dealing with numbers and symbols. So I instead
decided to create a symbols layer all on the right hand, and a numbers
layer all on the left hand. I also added a third layer which is
basically the movement layer I had before but with the symbols removed.