Collapse number & symbol layers together
I did this to simplify my layout for maintaining going forward but also
because if I am holding the modifier on one hand to get to a symbol it
allows me to access the numbers with the other hand without having to
switch modifier keys. I still have both hands setup with the modifier
key to bring this layer up as I plan to generally use the alternate
handed modifier approach. However, this supports that and gives me the
flexibility to access both symbols and numbers under the same modifier.

I also added a COMNT key to the numbers half of the symbols and numbers
layer so that I could simply use the modifier and press that one key to
do CMD + / for me as doing the scrunch to pull that off is not
comfortable.

I also added a four WINX keys to the motion layer. These are combos
specifically for interacting with emmetapp, the window manager. They
resize and place windows into the four window stacks I have configured
in emmetapp.