Enable home row mods
See https://precondition.github.io/home-row-mods for details. Basically
it sets up my home row to act as the modifier keys if held and act as
the normal letters when tapped.

This definitely makes the keyboard feel lagged because it waits until
keyup to register a keypress rather than keydown like it normally would.
Also, there can be gotchas with with the TAPPING_TERM, how long you have
to hold the key down for it to register as a modifier. So this will
likely be something that has to be tweaked as my speed evolves.

Also, IGNORE_MOD_TAP_INTERRUPT and TAPPING_FORCE_HOLD are critical for
handling some of those cases. See the QMK documentation for details on
these settings.

I will have to see how I like this and if I feel like it is worth the
delay feeling.