Experimented with the Combos feature
Experimented with the Combos feature but ran into similar issues to the
MT() home row modifiers approach even when using a thumb cluster key to
trigger the combos. Also the ~40 ms COMBO_TERM still impacts the speed
of normal typing which is noticable. So for now I have switched back to
my layer approach to see if I can train myself to not chord the combos
and instead rock into them from the layer modifier.